// src/Enums/EnumInstructions.ts
/**
 * Represents the type of instruction specified by the user.
 */
export enum EnumInstructions {
    None = 0,

    NextInLine = 1,

    AddUser = 2,

    Assign = 3,

    Help = 4,

    Changelog = 5,

    Remove = 6,

    RemoveWithUser = 7,

    Scoreboard = 8,

    OofStatus = 9,

    Random = 10,

    ChannelRandomness = 11
}